Fix undefined e element when submitting an empty search text
When a user wants to do a search and fills in an empty string (space) and then hits enter on the keyboard (especially on mobile devices with gboard this is an issue) a js error is thrown. This happens because this.superApply(e) is not passing the e correctly to the parent function. Instead of that a much simpler solution is implemented.