Avoid API Exceptions for empty product template ids
When retrieving the the related/upsell relations of a product for which the corresponding crosssell/upsell template id is empty, an `\Emico\Tweakwise\Exception\ApiException` is triggered with the following message:

    Featured products without template ID was requested.

This exception is caught later on and is written as an error to the `system.log`.

However, imho this scenario is not an actual error.

Also, this might give rise to lots of error messages, which pollutes the `system.log` (and services that process errors from these logs like Sentry and Elasticsearch).

This PR fixes these issues by adding an extra check to the Related and Upsell plugins before trying to perform a request to Tweakwise.